+++ projeto-manager/src/pages/SignUpPage/SignUpPage.js
@@ -6,10 +6,12 @@ import { DatePicker } from '@mui/lab'
 import { Body, Form, Main, TextSignUp } from "./styled";
 import { goBack, goToList } from "../../router/coordinator";
 import { useHistory } from "react-router-dom";
+import useForm from "../../hooks/useForm";
 
 
 const FormPage = () => {
     const history = useHistory()
+    const [form, onChange, clear] = useForm({ email: "", password: "" })
     const [value, setValue] = useState(null);
     const onSubmitForm = (event) => {
         event.preventDefault()
@@ -28,6 +30,7 @@ const FormPage = () => {
                         required
                         id="outlined-required"
                         label="Name"
+                        onChange={onChange}
 
                     />
 
@@ -36,18 +39,21 @@ const FormPage = () => {
                         id="outlined-required"
                         label="E-mail"
                         type='email'
+                        onChange={onChange}
                     />
                     <TextField
                         required
                         id="outlined-required"
                         label=" Phone (xx) xxxxx-xxxx"
                         type='tel'
+                        onChange={onChange}
 
                     />
                     <TextField
                         required
                         id="outlined-required"
                         label="Kind of Service"
+                        onChange={onChange}
                     />
                     <LocalizationProvider dateAdapter={AdapterDateFns}>
                         <DatePicker
@@ -59,10 +65,21 @@ const FormPage = () => {
                             }}
                             renderInput={(params) => <TextField {...params} />}
                         />
+                        <TextField
+                            required
+                            id="outlined-required"
+                            label="Password"
+                            type='password'
+                            value={form.password}
+                            name={'password'}
+                            onChange={onChange}
+                            onClick={() => goBack(history)}
+
+                        />
                     </LocalizationProvider>
 
                     <Button
-                    onClick={() => goToList(history)}
+                        onClick={() => goToList(history)}
                         color='secondary'
                         variant="outlined">
                         Sing Up
